refactor: atom family 추가
